requirements.txt
accept-paymob==1.0.0
asgiref==3.3.4
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
chroma-py==0.1.0.dev1
colour==0.1.5
cryptography==3.4.8
cycler==0.10.0
defusedxml==0.7.1
dj-database-url==0.5.0
dj-rest-auth==2.1.11
Django==3.2.4
django-allauth==0.45.0
django-crispy==1.8.1
django-crispy-forms==1.12.0
django-decouple==2.1
django-heroku==0.3.1
django-location-field==2.1.0
django-map-widgets==0.3.2
django-mapbox-location-field==1.7.3
django-rest-auth==0.9.5
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
geographiclib==1.52
geojson==2.5.0
geopy==2.2.0
gunicorn==20.1.0
idna==2.10
Jinja2==3.0.1
kiwisolver==1.3.2
leaflet==0.0.3
mapbox-location==1.0.3
mapboxgl==0.10.2
MarkupSafe==2.0.1
matplotlib==3.4.3
numpy==1.21.2
oauthlib==3.1.1
paymob-accept==0.1.1
Pillow==8.2.0
psycopg2==2.9.1
pycparser==2.20
PyJWT==2.0.0
pyparsing==2.4.7
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
six==1.16.0
sqlparse==0.4.1
twilio==6.60.0
urllib3==1.26.5
whitenoise==5.2.0